requirements.txt
# Core dependencies
fastmcp>=0.1.0
google-cloud-bigquery>=3.11.0
google-auth>=2.17.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
black>=23.7.0
flake8>=6.1.0
mypy>=1.4.1
types-PyYAML>=6.0.12
